Time: 3 ms (7.94%) | Memory: 32.1 MB (6.35%) - LeetSync

+++ 2238-a-number-after-a-double-reversal/a-number-after-a-double-reversal.cs
@@ -2,9 +2,9 @@ public class Solution {
     public bool IsSameAfterReversals(int num) {
         string one = Convert.ToString(num);
         string rstr = new string(one.Reverse().ToArray());
-        BigInteger bint = BigInteger.Parse(rstr);
+        int bint = int.Parse(rstr);
         string tstr = Convert.ToString(bint);
         string rtstr = new string(tstr.Reverse().ToArray());
-        return num==BigInteger.Parse(rtstr);
+        return num==int.Parse(rtstr);
     }
 }
